check-revng-conventions: fix black invocation
Before this commit, the black code formatter for python was invoked even
if the list of python files was empty, causing errors in that case.

This never happened in revng, but it happens in other revng projects
that do not have python files.

This commit fixes the script, skipping the invocation to black whenever
no python files is found.