MakeModelGEPPass: fix comparison of QualifiedTypes
This comparison was used to sort the best results among candidates for
creating ModelGEPs.

However the comparison itself was broken, i.e. you could have 2
different QualifiedTypes T1 and T2 such that T1 < T2 AND T2 < T1.
This caused broken sorting and non-deterministic selection of the best
ModelGEP substitution.

This commit fixes the comparison and as a result the selection of the
best candidate for the emission of ModelGEPs is not completely
deterministic.