Remove static asserts in tuple tree references
Tuple tree references use Root type as a pointer only so they can
operate with forward declarations. Before this commits there were static
asserts that prevented this use case, and thus required the headers
defining the root type to be included before this one to work correctly.

We drop those static asserts to make sure that all headers are parsable
on their own.