IDBImporter: Fix generation of empty structs and unions
There was a log message that we are creating typedef void* for
empty structs, but we have missed that piece of code. This
patch adds that logic.

In the case of unions, we completely ignored empty ones, and also
there was not handling if the type was generated by ordinal (as
in the case of structs). This patch also fixes it.