Remove from `Backedges` set retreating edges
During the creation of the exit dispatcher, take care of removing from
the `Backedges` set the additional backedge that is collapsed in case
two exiting backedges do target the same destination.

When this is not done, a ghost backedge is left inside the set, making
it not coherent with the state of the graph.

In addition, we add an assertion which ensures that during this phase of
the restructuring dummy nodes are present only as source of retreating
edges, and have one and only one successor.