LLVMContainerBase::cloneFiltered: fix dbg metadata
Before this commit the cloneFiltered method was duplicating debug
metadata on cloned functions, causing the Module not to verify after
cloning.
The reason is a nasty combination of how llvm::cloneModule and
llvm::Function::clearMetadata work internally:
- llvm::cloneModule already clones debug metadata, so the cloned
  functions have debug metadata
- llvm::Function::clearMetadata does not clear debug metadata
- so it's wrong to manually copy debug metadata from old functions to
  cloned functions

This commit fixes the problem by skipping the copy of the debug
metadata alltoghether, since they're already in place.